+++ Client.php
@@ -1596,6 +1596,29 @@ class Client
     }
 
     /**
+     * POST /api/v1/sbp/check-async
+     */
+    public function sbpCheckAsync(
+        string $phone,
+        ?string $bic = null,
+        ?string $bankName = null
+    ) {
+        return $this->exec('POST', 'sbp/check-async', [], $this->filterParams([
+            'phone' => $phone,
+            'bic' => $bic,
+            'bank_name' => $bankName,
+        ]));
+    }
+
+    /**
+     * GET /api/v1/sbp/get-check-async-status
+     */
+    public function sbpGetCheckAsyncStatus(string $id)
+    {
+        return $this->exec('POST', 'sbp/check-async/status', [], $this->filterParams(['id' => $id]));
+    }
+
+    /**
      * POST /api/v1/sbp/payment
      */
     public function sbpPayment(
@@ -1629,6 +1652,32 @@ class Client
         return $this->exec('POST', sprintf('clients/%s/check-sbp', $clientId), [], $this->filterParams([
             'bic' => $bic,
             'bank_name' => $bankName,
+        ]));
+    }
+
+    /**
+     * POST /api/v1/clients/{client_id}/check-sbp-async
+     */
+    public function sbpClientCheckAsync(
+        string $clientId,
+        ?string $bic = null,
+        ?string $bankName = null
+    ) {
+        return $this->exec('POST', sprintf('clients/%s/check-sbp-async', $clientId), [], $this->filterParams([
+            'bic' => $bic,
+            'bank_name' => $bankName,
+        ]));
+    }
+
+    /**
+     * GET /api/v1/clients/{client_id}/check-sbp-async/status
+     */
+    public function sbpGetClientCheckAsyncStatus(
+        string $clientId,
+        string $id
+    ) {
+        return $this->exec('POST', sprintf('clients/%s/check-sbp-async/status', $clientId), [], $this->filterParams([
+            'id' => $id,
         ]));
     }